requirements.txt
adbc_driver_manager==1.9.0
adbc_driver_postgresql==1.9.0
adbc_driver_sqlite==1.9.0
anywidget==0.9.21
AppKit==0.2.8
atheris==3.0.0
authlib==1.6.6
beautifulsoup4==4.14.3
boto3==1.42.19
botocore==1.42.19
brotli==1.2.0
brotlicffi==1.2.0.0
checks==0.2
choreographer==1.2.1
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==46.0.3
cudf==0.6.1.post1
cupy==13.6.0
cycler==0.12.1
Cython==3.2.3
dask==2025.12.0
dask_expr==2.0.0
defusedxml==0.7.1
docutils==0.22.4
duckdb==1.4.3
fastparquet==2025.12.0
filelock==3.20.1
Foundation==0.1.0a0.dev1
fqdn==1.5.1
fsspec==2025.12.0
geopandas==1.1.2
gitdb_speedups==0.1.0
graphviz==0.21
h2==4.3.0
HTMLParser==0.0.2
huggingface_hub==1.2.3
hypothesis==6.148.9
ibis==3.3.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipython==8.12.3
ipywidgets==8.1.8
isoduration==20.11.0
jnius==1.1.0
JPype1==1.6.0
jsonpath_ng==1.7.0
jsonpointer==3.0.0
keyring==25.7.0
lxml==6.0.2
matplotlib==3.10.8
mock==5.2.0
modin==0.37.1
mtrand==0.1
mypy==1.19.1
numexpr==2.14.1
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
polars==1.36.1
polars_st==0.4.3
psutil==7.2.1
py==1.11.0
pycares==5.0.1
pycurl==7.45.7
PyInstaller==6.17.0
pyodide==0.0.2
pyOpenSSL==25.3.0
pyperf==2.9.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.10.1
PySide6==6.10.1
pyspark==4.1.0
pytest==9.0.2
pytest_run_parallel==0.8.1
python_calamine==0.6.1
pyxlsb==1.0.10
PyYAML==6.0.3
qtpy==2.4.3
railroad==0.5.0
redis==7.1.0
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
rfc3987_syntax==1.1.0
s3fs==2025.12.0
scikit_learn==1.8.0
scipy==1.16.3
scipy_doctest==2.0.1
sets==0.3.2
setuptools==80.9.0
setuptools_scm==9.2.2
simplejson==3.20.2
skimage==0.0
sniffio==1.3.1
snowflake==1.10.0
Sphinx==9.1.0
SQLAlchemy==2.0.45
sqlframe==3.44.0
statsmodels==0.14.6
streamlit_pdf==2.0.1
sympy==1.14.0
tables==3.10.2
thread==2.0.6
threadpoolctl==3.6.0
traitlets==5.14.3
trio==0.29.0
uri_template==1.3.0
urllib3_secure_extra==0.1.0
uvloop==0.22.1
vegafusion==2.0.3
vincent==0.4.4
webcolors==25.10.0
xarray==2025.12.0
xlrd==2.0.2
xlsxwriter==3.2.9
xmlrpclib==1.0.1
